refactor: restructure color palette generator layout and routes
Promote the theme builder to the root route and move the legacy generator
to /old. Reorganize components into docs/, old/, shared/, and themebuilder/
folders; adjust contrast/about/layout routes; drop unused contrast helpers;
update e2e specs for the new routes.

Includes the render-time clamp fix in the root page (safeContrastIndex),
which depends on the promoted theme-builder code and so lands here.
