+++ apps/eds-color-palette-generator/src/components/docs/BellCurveVisualization.tsx
