+++ apps/eds-color-palette-generator/src/components/old/AnchorColorInput.tsx
