+++ apps/eds-color-palette-generator/src/components/old/ColorScale.tsx
