+++ apps/eds-color-palette-generator/src/components/old/ColorScalesHeader.tsx
