+++ apps/eds-color-palette-generator/src/components/old/DisplayOptionsPanel.tsx
