+++ apps/eds-color-palette-generator/src/components/old/GaussianParametersPanel.tsx
