+++ apps/eds-color-palette-generator/src/components/old/LightnessValueInputs.tsx
