+++ apps/eds-color-palette-generator/src/components/old/NameAndControls.tsx
