+++ apps/eds-color-palette-generator/src/components/old/QuickActionsPopover.tsx
