+++ apps/eds-color-palette-generator/src/components/shared/Badge.tsx
