+++ apps/eds-color-palette-generator/src/components/shared/ThemeToggle.tsx
